update and delete Employee

+++ employee-api/src/main/java/com/codeolate/employee/api/repositories/EmployeeRepository.java
@@ -1,12 +1,11 @@
 package com.codeolate.employee.api.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.codeolate.employee.api.entity.Employee;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
+public interface EmployeeRepository extends JpaRepository<Employee, Long>{
 
 }
